+- add patches.fixes/linux-post-2.6.3-20040221 +  ieee1394, mips, ppc64 pci hotplug, networking updates
suse-commit: 4029d49454692260137cc27c9f72eb41becb0b10

+++ arch/mips/mm/pgtable-64.c
@@ -64,7 +64,7 @@ void __init pagetable_init(void)
 
 	memset((void *)kptbl, 0, PAGE_SIZE << PGD_ORDER);
 	memset((void *)kpmdtbl, 0, PAGE_SIZE);
-	set_pgd(swapper_pg_dir, __pgd(kpmdtbl));
+	set_pgd(swapper_pg_dir, __pgd((unsigned long)kpmdtbl));
 
 	/*
 	 * The 64-bit kernel uses a flat pagetable for it's kernel mappings ...
